+++ Scripts/Enums/NetworkMode.cs
@@ -1,0 +1,9 @@
+namespace DSC.Network
+{
+    public enum NetworkMode
+    {
+        Host,
+        Client,
+        Server
+    }
+}