+++ CHANGELOG.md
@@ -1,1 +1,4 @@
-## [0.0.0-preview.4] - 2021-03-25
+## [0.0.0-preview.5] - 2021-03-25
+- Add NetworkMode enum.
+- Add DSC_Network_HUD script.
+- Add DSC_Event_NetworkMode script.